refactor: Lotto클래스 이름 변경
- Lotto -> LottoTicket
